+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 4/20/2018
 - [Added] server/config.js
 - [Modified] server/server.js, config port
+- [Modified] server/server.js, DB connect setup
 
 4/15/2018
 - 使用 Node.js 與 Angular 5，建立模仿Amazon的網站